requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
gunicorn
yubico_client
# ecdsa
# paramiko
# pycrypto
